requirements.txt
# Core Flask dependencies
Flask==2.3.3
Flask-SQLAlchemy==3.0.3
Flask-Caching==2.0.2
Werkzeug==2.3.7

# Document processing (CORE FUNCTIONALITY)
docxtpl==0.16.7  # Template processing
python-docx==1.1.0  # DOCX manipulation
python-dateutil==2.8.2
aspose-words  # PDF conversion with perfect fidelity

# Image processing
Pillow>=9.5.0

# Development/Monitoring
tqdm==4.65.0
pytz==2023.3